Frontend ticket delete policy

+++ app/Livewire/Frontend/Ticket/DeleteModal.php
@@ -15,11 +15,13 @@ class DeleteModal extends ModalComponent
     }
     public function render()
     {
+        $this->authorize('delete',Ticket::find($this->rowId));
         return view('livewire.frontend.ticket.delete-modal');
     }
 
     public function delete_ticket()
     {
+        $this->authorize('delete',Ticket::find($this->rowId));
         $ticket = Ticket::find($this->rowId);
         $this->authorize('delete',$ticket);
         $ticket->delete();
